Add cursorPos handling to new input system
Convert Camera to use it. Remove some old code paths.

I don't handle the removing of the handle when Camera is destroyed. I
really dislike having to keep track of these and write all that
boilerplate. It's improving, I got to delete all the cursorPos callbacks
that just forwarded the call. But if I had a way to automatically handle
the cleanup, this would start looking good