Make cleaning up input handlers easier
Using the Camera and cursorPos as an example, I keep track of the
pointer to the instance and use that as the key in the map. Then, when
the Camera destructor is called, it tells UIContext to get rid of all
handlers associated with that class. Now, I don't need to makeup dumb
fake ids and store them and pass them around to cleanup the handlers. :)