Minimalize Window even more
Move the quit keybinding to App, which makes Window keybinding-free,
which seems much more appropriate given that it's meant to be a lower
level thing. App and friends seem like a better place to think about
closing the application/game.

Moved the glfw callback bindings to UIContext constructor. It makes
sense to me that UIContext can be responsible for setting up the
callbacks it cares about. As part of that, it also sets the UserPointer
now, so this class definitely takes over glfw by a certain definition.

But I think that's all fine, and this has turned out to be pretty useful
in reducing the boilerplate I had before to handle input.