Store input from new world page and display it
I decided to add redux back and tried to use it again. I managed to
figure out a way that works using dynamic casting. It's a little ugly,
but way better than any of my other previous attempts. I think it's good
enough to actually try to use at least.

I'll probably move current page into it too.

I'm going to store the Store in UIContext for the foreseeable future. I
first thought that it might be the other way around, but UIContext is so
core in that it deals with inputs and window events that everything
needs, so I already always have to pass that around, whereas the redux
Store seems more specific, and it's so convenient to have saved in the
UIContext since everything already has that. So it should be easy to
experiment with, and if I decide to rip it out, I don't have to change a
million classes.

What this is also maybe telling me is that instead of having everything
in "UIContext" I should rename the all encompasing container that I use
as the thing I pass around. I could call it Game or Context, whatever.
It would have both a UIContext and a Store that consumers of the Context
could both use. Then adding or removing functionality doesn't require
changing a million constructors that all pass individual bits around.

Something else I want to fix is that I'm copying the state struct all
the time. This isn't javascript land where we have immutable data
structures, I don't think adding this copy does any good, so I'd rather
have Store pass around references to State to the reducers. I think that
would involve forking the library and changing the class template to use
STATE_T& instead of STATE_T where desired.

Also need to look into how I would compose reducers, handle multiple
actions. Should probably make sure that actually works, I'm pretty sure
it does though.