Center chunk boundaries on camera
Since the world is mostly empty, to prevent too much line noise being
generated, I only render a small area of chunks boundaries around the
camera.

This also introdues the chunkPosition helper. It's pretty generic, so I
added it as a static function to the Chunk struct. But right now it's
only displayed in the gui and used for positioning these lines.