Replace redux library with a customized version to return refs
In my quest to adapt the whole redux style to c++. Returning a mutable
ref is super useful because I can wire up values to ImGui widgets. Doing
that, I don't need to somehow detect a value change, dispatch an action,
and update the state value which would be insane to try to do for every
changeable thing in state. It breaks the whole unidirectional data flow
thing which is kiiiind of a large part of the point of redux. It would
definitely break some things that normally would be possible with a
redux design (storing and replaying events for debugging, etc).

So it's a bit of an unfortunate position, I'm going to keep
experimenting though and maybe there's a better way out there.

I could even go further with some other library changes, such as using
Action refs maybe? I'm not sure if I can cleanup that to maybe not use
shared pointer but it would look nicer if I could.