Add slider for scale to view realtime change
Finally, part of my vision is really being able to interact with the
scene using debug tools, and being able to change the scale in realtime
makes it possible to visualize the differences between 1.0 and 1/32.0f.

I still kind of want 1 unit block outlines to give more sense of scale
but I feel like it would be hard to do well. Seems really crowded if I
have the outlines in empty space as well, which means it probably needs
to wait until I start adding in real meshing and what not.