Simplify franken-redux implementation
I'm just getting rid of the reducers concept. In JS land, it's really
nice to have separate reducers that can be combined into a single
function, but that whole thing didn't really work in C++ land very well.
Maybe there's a way to make it work? But I just went with a single
struct, State, which will contain all the stuff, and Actions are
basically just functions on the state.