Generate perlin noise terrain
Yay finally something else to look at for a while.

It's easier to see if scale factor is increased, the additional blocks
in the same amount of space means the terrain is generated at a higher
resolution.

A normal map would be helpful though in lieu of actual block colors or
lighting or anything for now.