Replace State struct with entt library
I'm pretty excited about this. One of the things I put off was how I was
going to be able to modularize the state. It would be pretty
inconvenient to have to put everything in State as the program grows.

It has the Context variables helpers which I'm pretty much using
exclusively now since I only have World and a CurrentPage. The
CurrentPage wrapper around the renderer ptr is kind of annoying, but
I think it's better than just having the shared_ptr<Renderer> be the
Context type (if that would even work).