Get very rough save/load working
The thing I messed around with forever is trying to save entt context
vars. They are ignored by entt::snapshot, and I couldn't a way to add
them to the serialization, although I suspect it's possible.

So I'm just going to have my own entity to basically use as a context
var for the stuff I want persisted.

Doing this raised some other issues, like how much imgui callback widget
will fire when the value changes, even if the widget wasn't clicked to
make it happen. Right now it has the effect of actually regenerating the
world chunks when I load, so ironically I'm getting the right behavior
for all the wrong reasons.

Once I fix the imgui callback widget and the chunks stop regenerating by
accident, I need to actually serialize the chunks so the block data is
saved/reloaded.

Once that works, I can clean that up a bit, and I'm hoping that leaves
me with a good foundation for adding more.