Fix firing callback when it shouldn't
I think imgui components in general return a boolean value when they are
interacted with, just like ImGui::Button. So it turns out I didn't need
any of that code I wrote, so I learned a bit more about imgui there.

There's also a lot more ways to query interaction state which may be
useful in the future. See:
https://github.com/ocornut/imgui/issues/1875#issuecomment-396752570

Now that my deserialized world doesn't cause chunks to regen, the chunks
array stays empty. What should really happen is that the chunks are
serialized and deserialized and reappear when the world is loaded.