Fix bug where chunk index was calculated wrong
I was being so dense, the chunk index was totally wrong. I feel like I
just guessed at calculating it, because once I compared it to the
for-loops in the world generate action, it was clearly wrong.

X stays 0, then Y stays 0, then Z goes from 0 to length. I was counting
left to right, then back to front, but it actually was back to front
then left to right. (and up to down, but w/e height is 1 for now. I'm
pretty sure that's correct though since it's in the middle it's easier
to see)