Remove unnecessary method
This was a bad idea, the method signature is absurd, and I've already
had to put other imgui stuff directly in the method, so I think it
should just go in there.

I imagine doing this all over would be useful for general debuggin
purposes, so ideally what would be cool is some way to manage all of
that debug display stuff. Camera is another example of a class that just
has imgui stuff I want to display.

I don't like the current overlay solution, it's far to intricately
connected. I want to make it easy for a totally new class to just barf
out debug info to the screen, and I have some ideas for how.

It's just a matter of when, I still want to finish this ray cast work
even though I've come up with 10 other things I'd also like to fix or
do in the meantime.