Add imgui helpers for tiling windows
It's super simple, but it gets the job done for now.

It would be great if the way I managed the menu bar height wasn't hard
coded.

The hard coded total windows is also annoying.

Probably a bunch of stuff I'll realize later once I try added more
windows or extending this in some way...

But for now, it works.