Reorganize the project a bit
I wanted to make a proper directory for the rendering system. Before I
just had the sort of 'renderers', but I imagine that's something I'll
phase out and replace with an entt system and components.

But yeah a directory for the camera, the shader classes, the debug
drawing manager, all the various parts.